backend/requirements.txt
# Django Core
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=24.0

# Authentication
django-allauth>=65.0
django-axes>=8.0
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9
dj-database-url>=2.0

# Utils
python-dotenv>=1.0

# Production server
gunicorn>=21.0
whitenoise>=6.6
